Rename `Registers` to `BasicRegisters` and move to `ab-riscv-interpreter`
