+++ crates/execution/ab-riscv-interpreter/src/basic.rs
@@ -1,5 +1,8 @@
 //! Basic implementations of various interpreter traits
 
+#[cfg(test)]
+mod tests;
+
 use crate::{
     Address, CustomErrorPlaceholder, ExecutionError, FetchInstructionResult, InstructionFetcher,
     ProgramCounter, ProgramCounterError, VirtualMemory,
@@ -7,6 +10,60 @@ use crate::{
 use ab_riscv_primitives::prelude::*;
 use core::marker::PhantomData;
 use core::ops::ControlFlow;
+
+/// A basic set of RISC-V GPRs (General Purpose Registers)
+#[derive(Debug, Clone, Copy)]
+#[repr(align(16))]
+pub struct BasicRegisters<Reg>
+where
+    Reg: Register,
+    [(); Reg::N]:,
+{
+    regs: [Reg::Type; Reg::N],
+}
+
+impl<Reg> Default for BasicRegisters<Reg>
+where
+    Reg: Register,
+    [(); Reg::N]:,
+{
+    #[inline(always)]
+    fn default() -> Self {
+        Self {
+            regs: [Reg::Type::default(); Reg::N],
+        }
+    }
+}
+
+const impl<Reg> BasicRegisters<Reg>
+where
+    Reg: [const] Register,
+    [(); Reg::N]:,
+{
+    /// Read register value
+    #[inline(always)]
+    pub fn read(&self, reg: Reg) -> Reg::Type {
+        if reg == Reg::ZERO {
+            // Always zero
+            return Reg::Type::default();
+        }
+
+        // SAFETY: register offset is always within bounds
+        *unsafe { self.regs.get_unchecked(usize::from(reg.offset())) }
+    }
+
+    /// Write register value
+    #[inline(always)]
+    pub fn write(&mut self, reg: Reg, value: Reg::Type) {
+        if reg == Reg::ZERO {
+            // Writes are ignored
+            return;
+        }
+
+        // SAFETY: register offset is always within bounds
+        *unsafe { self.regs.get_unchecked_mut(usize::from(reg.offset())) } = value;
+    }
+}
 
 /// Basic instruction fetcher implementation.
 ///
