+++ crates/execution/ab-riscv-primitives/src/lib.rs
@@ -62,7 +62,6 @@
     const_convert,
     const_default,
     const_destruct,
-    const_index,
     const_ops,
     const_option_ops,
     const_trait_impl,
